requirements.txt
Flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
pypdf
python-docx
pytesseract
Pillow